initial commit of mix phx.new panic --database sqlite3
this is the beginning of the v3 re-write, which will:

- upgrade to Phoenix v1.7 (rc2)

- remove PETAL (which was great, but the tailwind and core components stuff in
  Phoenix v1.7 kinda makes it obsolete, and it's nicer to stick to plain phoenix)

- use a proper FSM to manage the running of each network (not the janky "do it
  all in the liveview process and hope it doesn't get disconnected or no-one
  else connects simultaneously" approach from v1/2)

- move to sqlite (from postgres) - no HA or anything fancy at this stage, just a
  single volume running on fly on the same machine as the application code (in
  future we might do a cool liteFS thing, but just waiting until it's a bit more
  mature)

- move to a proper "sign up, create an account" thing, with per-account API
  creds, rather than the "only works for us, creds in env vars" thing we've got
  going now

and much more. hopefully this goes ok. if not, git reset --hard panic-v2

note: for those playing along at home, v1 was the initial version I hacked
together for the _Creative AI symposium_ at UNSW in July '22, v2 was the version
displayed at the _Australian Cybernetic_ exhibition in Dec '23, and v3
will (hopefully) be the version that's installed properly in Birch Lv3 foyer, as
well as being easier to take on the road.