all API calls (and associated stuff) now need a user arg
this is because it needs to pull the user's API token from the DB

honestly having to pass it in explicitly in all these cases is nice and
pure(ish)-functional-y, but is a pain in the arse. Might investigate whether
there's a nicer way to e.g. store the user as a module attribute or something.