change :new_prediction event to :new_prediction? (can fail)
The compiler still gives a warning when the transition fails, so I'm not sure
what that's about. But if I try to add the `waiting --> |new_prediction|
waiting` transition then the compiler complains about unhandled/ambiguous
transitions. So I'm not sure what's going on there, will investigate ASAP, so
that I can at least file a good issue.