add vestaboards array to %Network{}
it's a bit messy for now, and relies on some hacky Jason.decode stuff, but
it'll do for the MVP

from the discussion on today's Elixir Podcast it looks like the forms stuff is a
bit in flux, really - I'd love to see some examples of more complex forms using
this new FormData stuff. Perhaps I'll ask a question on the forum.