WIP: big simplification of %Predictions{}
gone back to the generator, ditching all the "genesis/next" machinery

there are some bits of the old code that I will cherry-pick back in, but in
general Chris McCord's commend on that podcast that "Changesets shouldn't hang
around" has been a lightbulb moment that I was doing it wrong. this new thing
will be simpler, more robust and just generally better.

this will take a little while to finalise the surgery on, but I feel like this
is the right change. and if not, I can always come back to the previous commit :shrug: