update model testing
the "mock the full platform responses" plan hit a snag in that model.invoke for
a Replicate model is actually _heaps_ of requests (due to the create-and-poll
approach their API uses). Trying to mock this series of requests seems like it'd
be flaky and not a great option, so I'll :skip all replicate tests for now. That
approach is ok for OpenAI, though