add Styler
this is a big commit diff-wise (because I just did a mix format), but the tests still pass
and I'm *pretty sure* that this codebase doesn't have any of the idioms that Styler breaks

Anyway, if it's a disaster I can revert. But I like the idea of not having to worry about
that stuff and just have Styler handle it for me.